independent multi-fidelity searcher
