+++ syne_tune/optimizer/schedulers/searchers/bore/__init__.py
@@ -4,14 +4,10 @@ import logging
 
 try:
     from syne_tune.optimizer.schedulers.searchers.bore.bore import Bore  # noqa: F401
-    from syne_tune.optimizer.schedulers.searchers.bore.multi_fidelity_bore import (  # noqa: F401
-        MultiFidelityBore,
-    )
 
     __all__.extend(
         [
             "Bore",
-            "MultiFidelityBore",
         ]
     )
 except ImportError as e:
