+++ syne_tune/optimizer/schedulers/searchers/multi_fidelity_searcher.py
@@ -1,24 +1,25 @@
 import logging
+from collections import OrderedDict
 
-from typing import Dict, Any
+from typing import Dict, Any, Optional, List, Union
 
 from syne_tune.optimizer.schedulers.searchers.searcher import BaseSearcher
+from syne_tune.optimizer.schedulers.searchers.single_objective_searcher import SingleObjectiveBaseSearcher
+from syne_tune.optimizer.schedulers.searchers.searcher_factory import searcher_cls_dict
 
 logger = logging.getLogger(__name__)
 
 
-class MultiFidelityBaseSearcher(BaseSearcher):
+class IndependentMultiFidelitySearcher(BaseSearcher):
     """
-    Base class of searchers, which are components of schedulers responsible for
-    implementing :meth:`get_config`.
+    Searcher for the multi-fidelity setting which fits independent models for each
+    resource level as as proposed by Falkner et al.
+
+    | BOHB: Robust and Efficient Hyperparameter Optimization at Scale
+    | S. Falkner and A. Klein and F. Hutter
+    | Proceedings of the 35th International Conference on Machine Learning
 
     # TODO: Update docstrings
-    .. note::
-       This is an abstract base class. In order to implement a new searcher, try to
-       start from
-       :class:`~syne_tune.optimizer.scheduler.searcher.StochasticAndFilterDuplicatesSearcher`
-       or :class:`~syne_tune.optimizer.scheduler.searcher.StochasticSearcher`,
-       which implement generally useful properties.
 
     :param config_space: Configuration space
     :param points_to_evaluate: List of configurations to be evaluated
@@ -29,6 +30,68 @@ class MultiFidelityBaseSearcher(BaseSearcher):
         determined by the midpoint heuristic. If ``[]`` (empty list), no initial
         configurations are specified.
     """
+
+
+    def __init__(
+            self,
+            config_space: Dict[str, Any],
+            searcher_cls: Optional[Union[str, SingleObjectiveBaseSearcher]] = "kde",
+            points_to_evaluate: Optional[List[dict]] = None,
+            random_seed: Optional[int] = None,
+            searcher_kwargs: dict[str, Any] = None
+    ):
+        super().__init__(
+            config_space,
+            points_to_evaluate=points_to_evaluate,
+            random_seed=random_seed,
+        )
+
+        if searcher_kwargs is None:
+            self.searcher_kwargs = dict()
+        else:
+            self.searcher_kwargs = searcher_kwargs
+
+        if isinstance(searcher_cls, str):
+            assert searcher_cls in searcher_cls_dict
+            self.searcher_cls = searcher_cls_dict.get(searcher_cls)
+        else:
+            self.searcher_cls = searcher_cls
+        self.models = OrderedDict()
+
+
+    def initialize_model(self):
+        return self.searcher_cls(
+            config_space=self.config_space,
+            random_seed=self.random_seed,
+            **self.searcher_kwargs,
+        )
+
+    def suggest(self, **kwargs) -> Optional[Dict[str, Any]]:
+        """Suggest a new configuration.
+
+        Note: Query :meth:`_next_points_to_evaluate` for initial configs to return
+        first.
+
+        :param kwargs: Extra information may be passed from scheduler to
+            searcher
+        :return: New configuration. The searcher may return None if a new
+            configuration cannot be suggested. In this case, the tuning will
+            stop. This happens if searchers never suggest the same config more
+            than once, and all configs in the (finite) search space are
+            exhausted.
+        """
+        suggestion = self._next_points_to_evaluate()
+        if suggestion is not None:
+            return suggestion
+
+        if len(self.models) == 0:
+            return {
+                k: v.sample() if hasattr(v, "sample") else v
+                for k, v in self.config_space.items()
+            }
+
+        highest_observed_resource = next(reversed(self.models))
+        return self.models[highest_observed_resource].suggest()
 
     def on_trial_result(
         self,
@@ -51,7 +114,12 @@ class MultiFidelityBaseSearcher(BaseSearcher):
         :param config: See :meth:`~syne_tune.optimizer.schedulers.TrialScheduler.on_trial_result`
         :param metric: See :meth:`~syne_tune.optimizer.schedulers.TrialScheduler.on_trial_result`
         """
-        return
+        if resource_level not in self.models:
+            self.models[resource_level] = self.initialize_model()
+
+        self.models[resource_level].on_trial_complete(
+            trial_id=trial_id, config=config, metric=metric
+        )
 
     def on_trial_complete(
         self,
@@ -74,4 +142,10 @@ class MultiFidelityBaseSearcher(BaseSearcher):
         :param config: See :meth:`~syne_tune.optimizer.schedulers.TrialScheduler.on_trial_result`
         :param metric: See :meth:`~syne_tune.optimizer.schedulers.TrialScheduler.on_trial_result`
         """
-        return
+
+        if resource_level not in self.models:
+            self.models[resource_level] = self.initialize_model()
+
+        self.models[resource_level].on_trial_complete(
+            trial_id=trial_id, config=config, metric=metric
+        )
