+++ syne_tune/optimizer/schedulers/searchers/random_searcher.py
@@ -1,9 +1,6 @@
 import logging
 from typing import Optional, List, Dict, Any
 
-from syne_tune.optimizer.schedulers.searchers.multi_fidelity_searcher import (
-    MultiFidelityBaseSearcher,
-)
 from syne_tune.optimizer.schedulers.searchers.searcher import BaseSearcher
 from syne_tune.optimizer.schedulers.searchers.single_objective_searcher import (
     SingleObjectiveBaseSearcher,
@@ -48,35 +45,6 @@ class RandomSearcher(SingleObjectiveBaseSearcher):
 
 
 class MultiObjectiveRandomSearcher(BaseSearcher):
-    """
-    Searcher which randomly samples configurations to try next.
-    """
-
-    def __init__(
-        self,
-        config_space: Dict[str, Any],
-        points_to_evaluate: Optional[List[dict]] = None,
-        random_seed: int = None,
-    ):
-        super().__init__(
-            config_space, points_to_evaluate=points_to_evaluate, random_seed=random_seed
-        )
-
-    def suggest(self) -> Optional[dict]:
-        """Sample a new configuration at random
-
-        If ``allow_duplicates == False``, this is done without replacement, so
-        previously returned configs are not suggested again.
-
-        :return: New configuration, or None
-        """
-        new_config = self._next_points_to_evaluate()
-        if new_config is None:
-            new_config = sample_random_config(self.config_space)
-        return new_config
-
-
-class MultiFidelityRandomSearcher(MultiFidelityBaseSearcher):
     """
     Searcher which randomly samples configurations to try next.
     """
